@@ -23,99 +23,98 @@ import vn.tpsc.it4u.model.enums.UserType;
 
 /**
  * CustomUserDetails
- * Spring Security will use the information stored in the this class' object to perform authentication and authorization
+ * Spring Security will use the information stored in the this class' object to
+ * perform authentication and authorization
  */
 @Data
 @AllArgsConstructor
 @EqualsAndHashCode
 public class CustomUserDetails implements UserDetails {
 
-    /**
-     *
-     */
-    private static final long serialVersionUID = 1L;
+	/**
+	 *
+	 */
+	private static final long serialVersionUID = 1L;
 
-    @Id
-    private Long id;
+	@Id
+	private Long id;
 
-    private String name;
+	private String name;
 
-    private String username;
+	private String username;
 
-    @JsonIgnore
-    private String password;
+	@JsonIgnore
+	private String password;
 
-    @JsonIgnore
-    private String email;
+	@JsonIgnore
+	private String email;
 
-    private String avatar;
+	private String avatar;
 
-    private Gender gender;
+	private Gender gender;
 
-    private UserType type;
+	private UserType type;
 
-    private UserStatus status;
+	private UserStatus status;
 
-    private String sitename;
+	private String sitename;
 
-    private Set<Role> roles;
+	private Set<Role> roles;
 
-    private Collection<? extends GrantedAuthority> authorities;
+	private Collection<? extends GrantedAuthority> authorities;
 
-    public static CustomUserDetails create(User user) {
-        List<GrantedAuthority> authorities = user.getRoles().stream().map(role ->
-                new SimpleGrantedAuthority(role.getName().name())
-        ).collect(Collectors.toList());
+	public static CustomUserDetails create(User user) {
+		List<GrantedAuthority> authorities = user.getRoles().stream()
+				.map(role -> new SimpleGrantedAuthority(role.getName().name())).collect(Collectors.toList());
 
-        return new CustomUserDetails(
-                user.getId(),
-                user.getName(),
-                user.getUsername(),
-                user.getPassword(),                
-                user.getEmail(),
-                user.getAvatar(),
-                user.getGender(),
-                user.getType(),
-                user.getStatus(),
-                user.getSitename(),
-                user.getRoles(),
-                authorities
-        );
-    }
+		return new CustomUserDetails(
+				user.getId(),
+				user.getName(),
+				user.getUsername(),
+				user.getPassword(),
+				user.getEmail(),
+				user.getAvatar(),
+				user.getGender(),
+				user.getType(),
+				user.getStatus(),
+				user.getSitename(),
+				user.getRoles(),
+				authorities);
+	}
 
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return authorities;
-    }
+	@Override
+	public Collection<? extends GrantedAuthority> getAuthorities() {
+		return authorities;
+	}
 
-    @Override
-    public String getPassword() {
-        return password;
-    }
+	@Override
+	public String getPassword() {
+		return password;
+	}
 
-    @Override
-    public String getUsername() {
-        return username;
-    }
+	@Override
+	public String getUsername() {
+		return username;
+	}
 
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
+	@Override
+	public boolean isAccountNonExpired() {
+		return true;
+	}
 
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
+	@Override
+	public boolean isAccountNonLocked() {
+		return true;
+	}
 
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
+	@Override
+	public boolean isCredentialsNonExpired() {
+		return true;
+	}
 
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-    
+	@Override
+	public boolean isEnabled() {
+		return true;
+	}
+
 }
