@@ -17,29 +17,29 @@ import java.util.Optional;
  */
 @Configuration
 @EnableJpaAuditing
-//@EnableMongoAuditing
+// @EnableMongoAuditing
 public class AuditingConfig {
 
-    @Bean
-    public AuditorAware<Long> auditorProvider() {
-        return new SpringSecurityAuditAwareImpl();
-    }
+	@Bean
+	public AuditorAware<Long> auditorProvider() {
+		return new SpringSecurityAuditAwareImpl();
+	}
 }
 
 class SpringSecurityAuditAwareImpl implements AuditorAware<Long> {
 
-    @Override
-    public Optional<Long> getCurrentAuditor() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+	@Override
+	public Optional<Long> getCurrentAuditor() {
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
-        if (authentication == null ||
-                !authentication.isAuthenticated() ||
-                authentication instanceof AnonymousAuthenticationToken) {
-            return Optional.empty();
-        }
+		if (authentication == null ||
+				!authentication.isAuthenticated() ||
+				authentication instanceof AnonymousAuthenticationToken) {
+			return Optional.empty();
+		}
 
-        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+		CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
 
-        return Optional.ofNullable(userDetails.getId());
-    }
+		return Optional.ofNullable(userDetails.getId());
+	}
 }
