@@ -9,16 +9,16 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(HttpStatus.BAD_REQUEST)
 public class BadRequestException extends RuntimeException {
 
-    /**
-     *
-     */
-    private static final long serialVersionUID = 1L;
+	/**
+	 *
+	 */
+	private static final long serialVersionUID = 1L;
 
-    public BadRequestException(String message) {
-        super(message);
-    }
+	public BadRequestException(String message) {
+		super(message);
+	}
 
-    public BadRequestException(String message, Throwable cause) {
-        super(message, cause);
-    }
+	public BadRequestException(String message, Throwable cause) {
+		super(message, cause);
+	}
 }
