@@ -10,22 +10,19 @@ import lombok.Setter;
 import javax.persistence.MappedSuperclass;
 
 @MappedSuperclass
-@JsonIgnoreProperties(
-        value = {"createdBy", "updatedBy"},
-        allowGetters = true
-)
+@JsonIgnoreProperties(value = { "createdBy", "updatedBy" }, allowGetters = true)
 @Getter
 @Setter
 public abstract class UserDateAudit extends DateAudit {
 
-    /**
-     *
-     */
-    private static final long serialVersionUID = 1L;
+	/**
+	 *
+	 */
+	private static final long serialVersionUID = 1L;
 
-    @CreatedBy
-    private Long createdBy;
+	@CreatedBy
+	private Long createdBy;
 
-    @LastModifiedBy
-    private Long updatedBy;
+	@LastModifiedBy
+	private Long updatedBy;
 }
