@@ -1,6 +1,6 @@
 package vn.tpsc.it4u.model.enums;
 
 public enum Gender {
-    Male,
-    Female
+	Male,
+	Female
 }
