@@ -1,7 +1,7 @@
 package vn.tpsc.it4u.model.enums;
 
 public enum RoleName {
-    ROLE_ADMIN,
-    ROLE_USER,
-    ROLE_MANAGER
+	ROLE_ADMIN,
+	ROLE_USER,
+	ROLE_MANAGER
 }
