@@ -1,7 +1,7 @@
 package vn.tpsc.it4u.model.enums;
 
 public enum UserStatus {
-    Active,
-    Block,
-    Deleted
+	Active,
+	Block,
+	Deleted
 }
