@@ -8,7 +8,7 @@ import lombok.Data;
 @Data
 public class ApiResponse {
     private int code;
-    
+
     private String message;
 
     private Object data;
