@@ -8,10 +8,10 @@ import lombok.Data;
 @Data
 public class JwtAuthenticationResponse {
 
-    private String accessToken;
-    private String tokenType = "Bearer";
+	private String accessToken;
+	private String tokenType = "Bearer";
 
-    public JwtAuthenticationResponse(String accessToken) {
-        this.accessToken = accessToken;
-    }
+	public JwtAuthenticationResponse(String accessToken) {
+		this.accessToken = accessToken;
+	}
 }
