@@ -12,9 +12,9 @@ import lombok.Data;
 @AllArgsConstructor
 public class LoginRequest {
 
-    @NotBlank
-    private String usernameOrEmail;
+	@NotBlank
+	private String usernameOrEmail;
 
-    @NotBlank
-    private String password;
+	@NotBlank
+	private String password;
 }
