@@ -13,10 +13,10 @@ import lombok.Setter;
 @Setter
 public class PagedResponse<T> {
 
-    private List<T> content;
-    private int page;
-    private int size;
-    private long totalElements;
-    private int totalPages;
-    private boolean last;
+	private List<T> content;
+	private int page;
+	private int size;
+	private long totalElements;
+	private int totalPages;
+	private boolean last;
 }
