@@ -15,28 +15,28 @@ import vn.tpsc.it4u.model.enums.UserType;
 @Data
 public class SignUpRequest {
 
-    @NotBlank
-    @Size(min = 4, max = 40)
-    private String name;
+	@NotBlank
+	@Size(min = 4, max = 40)
+	private String name;
 
-    @NotBlank
-    @Size(min = 3, max = 15)
-    private String username;
+	@NotBlank
+	@Size(min = 3, max = 15)
+	private String username;
 
-    @NotBlank
-    @Size(max = 40)
-    @Email
-    private String email;
+	@NotBlank
+	@Size(max = 40)
+	@Email
+	private String email;
 
-    @NotBlank
-    @Size(min = 6, max = 20)
-    private String password;
+	@NotBlank
+	@Size(min = 6, max = 20)
+	private String password;
 
-    private Gender gender;
+	private Gender gender;
 
-    private UserType type;
+	private UserType type;
 
-    private UserStatus status;
+	private UserStatus status;
 
-    private String sitename;
+	private String sitename;
 }
