@@ -13,15 +13,15 @@ import vn.tpsc.it4u.model.User;
  */
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-    
-    Optional<User> findByUsernameOrEmail(String username, String email);
+	Optional<User> findByEmail(String email);
 
-    List<User> findByIdIn(List<Long> userIds);
+	Optional<User> findByUsernameOrEmail(String username, String email);
 
-    Optional<User> findByUsername(String username);
+	List<User> findByIdIn(List<Long> userIds);
 
-    Boolean existsByUsername(String username);
+	Optional<User> findByUsername(String username);
 
-    Boolean existsByEmail(String email);
+	Boolean existsByUsername(String username);
+
+	Boolean existsByEmail(String email);
 }
