@@ -18,16 +18,16 @@ import org.springframework.stereotype.Component;
 @Component
 public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
 
-    private static final Logger logger = LoggerFactory.getLogger(CustomAuthenticationEntryPoint.class);
+	private static final Logger logger = LoggerFactory.getLogger(CustomAuthenticationEntryPoint.class);
 
-    @Override
-    //This method is called whenever an exception is thrown due to an unauthenticated user trying to access a resource that requires authentication
-    public void commence(HttpServletRequest request, HttpServletResponse response,
-            AuthenticationException e) throws IOException, ServletException {
+	@Override
+	// This method is called whenever an exception is thrown due to an
+	// unauthenticated user trying to access a resource that requires authentication
+	public void commence(HttpServletRequest request, HttpServletResponse response,
+			AuthenticationException e) throws IOException, ServletException {
 
-        logger.error("Responding with unauthorized error. Message - {}", e.getMessage());
-        response.sendError(HttpServletResponse.SC_UNAUTHORIZED, e.getMessage());
-    }
+		logger.error("Responding with unauthorized error. Message - {}", e.getMessage());
+		response.sendError(HttpServletResponse.SC_UNAUTHORIZED, e.getMessage());
+	}
 
-    
 }
