@@ -16,29 +16,26 @@ import vn.tpsc.it4u.repository.UserRepository;
 @Service
 public class CustomUserDetailsService implements UserDetailsService {
 
-    @Autowired
-    UserRepository userRepository;
+	@Autowired
+	UserRepository userRepository;
 
-    @Override
-    @Transactional
-    // This method is used by Spring Security
-    public UserDetails loadUserByUsername(String usernameOrEmail) throws UsernameNotFoundException {
-        // Let people login with either username or email
-        User user = userRepository.findByUsernameOrEmail(usernameOrEmail, usernameOrEmail)
-                .orElseThrow(() -> 
-                        new UsernameNotFoundException("User not found with username or email : " + usernameOrEmail)
-        );
+	@Override
+	@Transactional
+	// This method is used by Spring Security
+	public UserDetails loadUserByUsername(String usernameOrEmail) throws UsernameNotFoundException {
+		// Let people login with either username or email
+		User user = userRepository.findByUsernameOrEmail(usernameOrEmail, usernameOrEmail)
+				.orElseThrow(() -> new UsernameNotFoundException("User not found with username or email : " + usernameOrEmail));
 
-        return CustomUserDetails.create(user);
-    }
+		return CustomUserDetails.create(user);
+	}
 
-    // This method is used by JWTAuthenticationFilter
-    @Transactional
-    public UserDetails loadUserById(Long id) {
-        User user = userRepository.findById(id).orElseThrow(
-            () -> new UsernameNotFoundException("User not found with id : " + id)
-        );
+	// This method is used by JWTAuthenticationFilter
+	@Transactional
+	public UserDetails loadUserById(Long id) {
+		User user = userRepository.findById(id).orElseThrow(
+				() -> new UsernameNotFoundException("User not found with id : " + id));
 
-        return CustomUserDetails.create(user);
-    }
+		return CustomUserDetails.create(user);
+	}
 }
