@@ -14,48 +14,48 @@ import vn.tpsc.it4u.payload.ApiResponse;
 @Component
 public class ApiResponseUtils {
 
-    @Autowired
-    private MessageSource messageSource;
-    
-    public ApiResponse success(int code, Locale locale) {
-        ApiResponse response = new ApiResponse();
-        response.setCode(code);
-        response.setMessage(messageSource.getMessage(String.valueOf(code), null, "Success", locale));
+	@Autowired
+	private MessageSource messageSource;
 
-        return response;  
-    }
+	public ApiResponse success(int code, Locale locale) {
+		ApiResponse response = new ApiResponse();
+		response.setCode(code);
+		response.setMessage(messageSource.getMessage(String.valueOf(code), null, "Success", locale));
 
-    public ApiResponse success(String message) {
-        ApiResponse response = new ApiResponse();
-        response.setCode(1000);
-        response.setMessage(message);
+		return response;
+	}
 
-        return response;  
-    }
+	public ApiResponse success(String message) {
+		ApiResponse response = new ApiResponse();
+		response.setCode(1000);
+		response.setMessage(message);
 
-    public ApiResponse success(Object data, Locale locale) {
-        ApiResponse response = new ApiResponse();
-        response.setCode(1000);
-        response.setMessage(messageSource.getMessage(String.valueOf(1000), null, "Success", locale));
-        response.setData(data);
+		return response;
+	}
 
-        return response;        
-    }
+	public ApiResponse success(Object data, Locale locale) {
+		ApiResponse response = new ApiResponse();
+		response.setCode(1000);
+		response.setMessage(messageSource.getMessage(String.valueOf(1000), null, "Success", locale));
+		response.setData(data);
 
-    public ApiResponse error(int code, Locale locale) {
-        ApiResponse response = new ApiResponse();
-        response.setCode(code);
-        response.setMessage(messageSource.getMessage(String.valueOf(code), null, "Error", locale));
+		return response;
+	}
 
-        return response;        
-    }
+	public ApiResponse error(int code, Locale locale) {
+		ApiResponse response = new ApiResponse();
+		response.setCode(code);
+		response.setMessage(messageSource.getMessage(String.valueOf(code), null, "Error", locale));
 
-    public ApiResponse error(int code, Object errors, Locale locale) {
-        ApiResponse response = new ApiResponse();
-        response.setCode(code);
-        response.setMessage(messageSource.getMessage(String.valueOf(code), null, "Error",locale));
-        response.setErrors(errors);
+		return response;
+	}
 
-        return response;        
-    }
+	public ApiResponse error(int code, Object errors, Locale locale) {
+		ApiResponse response = new ApiResponse();
+		response.setCode(code);
+		response.setMessage(messageSource.getMessage(String.valueOf(code), null, "Error", locale));
+		response.setErrors(errors);
+
+		return response;
+	}
 }
