@@ -1,8 +1,8 @@
 package vn.tpsc.it4u.util;
 
 public interface AppConstants {
-    String DEFAULT_PAGE_NUMBER = "0";
-    String DEFAULT_PAGE_SIZE = "30";
+	String DEFAULT_PAGE_NUMBER = "0";
+	String DEFAULT_PAGE_SIZE = "30";
 
-    int MAX_PAGE_SIZE = 50;
+	int MAX_PAGE_SIZE = 50;
 }
