@@ -4,8 +4,7 @@ package vn.tpsc.it4u.util;
  * ObjectExtension
  */
 public class StringUtils {
-
-    public static Boolean isNullorEmpty(String object) {
+	public static Boolean isNullorEmpty(String object) {
 		return object == null || object.isEmpty();
 	}
 }
