@@ -7,13 +7,13 @@ import javax.net.ssl.X509TrustManager;
 
 public class TrustAnyTrustManager implements X509TrustManager {
 
-public void checkClientTrusted(X509Certificate[] chain, String authType) throws CertificateException {
-}
+	public void checkClientTrusted(X509Certificate[] chain, String authType) throws CertificateException {
+	}
 
-public void checkServerTrusted(X509Certificate[] chain, String authType) throws CertificateException {
-}
+	public void checkServerTrusted(X509Certificate[] chain, String authType) throws CertificateException {
+	}
 
-public X509Certificate[] getAcceptedIssuers() {
-return new X509Certificate[] {};
-}
+	public X509Certificate[] getAcceptedIssuers() {
+		return new X509Certificate[] {};
+	}
 }
